Create mapper class.
