+++ features/continue-watching/presenter/src/commonTest/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/FakeContinueWatchingPresenterBuilder.kt
@@ -131,7 +131,7 @@ class FakeContinueWatchingPresenterBuilder {
         syncObserver = syncObserver,
         traktAuthRepository = fakeTraktAuthRepository,
         errorToStringMapper = ErrorToStringMapper { it.message ?: "Test error" },
-        localizer = localizer,
+        mapper = ContinueWatchingMapper(localizer),
         logger = fakeLogger,
     )
 }
